pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "otool-esptool-ui"
version = "0.1.1"
description = "PyQt6 GUI for mass-detecting and flashing ESP devices with esptool"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "ONEXS"}]
license = {file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "PyQt6>=6.0.0",
    "esptool>=4.9",
    "bitstring>=3.1.6,!=4.2.0",
    "cryptography>=43.0.0",
    "pyserial>=3.3",
    "reedsolo>=1.5.3,<1.8",
    "PyYAML>=5.1",
    "intelhex",
    "rich_click<2",
    "click<9",
]

[project.scripts]
otool-esptool-ui = "otool_esptool_ui:main"

[tool.setuptools]
packages = ["otool_esptool_ui", "otool_esptool_ui.src"]

[tool.setuptools.package-dir]
otool_esptool_ui = "."